+++ api/app/controllers/old/attachment.rb
@@ -1,0 +1,12 @@
+require "digest/sha2"
+
+require "sinatra/activerecord_helpers"
+require "sinatra/json_helpers"
+require_relative "../services/account_service"
+
+# ファイルアップロード
+class AttachmentRoutes < Sinatra::Base
+  helpers Sinatra::ActiveRecordHelpers
+  helpers Sinatra::JSONHelpers
+  helpers Sinatra::AccountServiceHelpers
+end
