+++ api/app/controllers/old/contest.rb
@@ -1,0 +1,6 @@
+require "sinatra/json_helpers"
+
+# 大会に関する情報を提供する
+
+class ContestRoutes < Sinatra::Base
+end
