+++ api/app/controllers/old/notice.rb
@@ -1,0 +1,13 @@
+require "sinatra/activerecord_helpers"
+require "sinatra/json_helpers"
+require_relative "../services/account_service"
+require_relative "../services/nested_entity"
+require_relative "../services/notification_service"
+
+class NoticeRoutes < Sinatra::Base
+  helpers Sinatra::ActiveRecordHelpers
+  helpers Sinatra::NestedEntityHelpers
+  helpers Sinatra::JSONHelpers
+  helpers Sinatra::AccountServiceHelpers
+  helpers Sinatra::NotificationService
+end
