+++ api/app/controllers/old/problem_group.rb
@@ -1,0 +1,10 @@
+require "sinatra/activerecord_helpers"
+require_relative "../services/account_service"
+require_relative "../services/nested_entity"
+
+class ProblemGroupRoutes < Sinatra::Base
+  helpers Sinatra::ActiveRecordHelpers
+  helpers Sinatra::NestedEntityHelpers
+  helpers Sinatra::JSONHelpers
+  helpers Sinatra::AccountServiceHelpers
+end
