+++ api/app/controllers/old/scoreboard.rb
@@ -1,0 +1,17 @@
+require "sinatra/activerecord_helpers"
+require "sinatra/json_helpers"
+require 'sinatra/competition_helpers'
+require_relative "../services/account_service"
+
+# 2日目の午後開始前まで確認可能
+# - 自分の順位と得点
+# - 上位3チームの得点とチーム名
+# - 自分の1つ上のチームの得点
+# - 各問題が何チームに解かれたか
+
+class ScoreBoardRoutes < Sinatra::Base
+  helpers Sinatra::ActiveRecordHelpers
+  helpers Sinatra::JSONHelpers
+  helpers Sinatra::AccountServiceHelpers
+  helpers Sinatra::CompetitionHelpers
+end
